Исправил проблему с превышением размера int при подсчете веса

+++ src/main/java/com/example/autoassignee/choose/assignee/Candidate.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class Candidate implements Comparable {
     private Reviewer reviewer;
-    private Integer weight;
+    private Long weight;
 
     @Override
     public boolean equals(Object obj) {
